# src/assets/json/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Application Name",
    "version": "1.0.0",
    "title": "Usus CI / CD API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "leander.gebhardt@student.htw-berlin.de"
    }
  },
  "host": "ba-2021.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/user/{username}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Gets specific user with username",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username of User",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error retrieving User"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updates user with username",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The username for login",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Project object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Invalid username/password supplied"
          },
          "500": {
            "description": "Error updating User"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Deletes User with Id",
        "description": "",
        "operationId": "logoutUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The users username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": {
          "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "full_name": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  },
...
}